% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clients_utils.R
\name{client_by_dpt}
\alias{client_by_dpt}
\title{Customer by departement}
\usage{
client_by_dpt(x, priority, age_class)
}
\arguments{
\item{x}{database}

\item{priority}{priority}

\item{age_class}{age_class}
}
\description{
Customer by departement
}
