% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tickets.R
\name{mod_ticketsui}
\alias{mod_ticketsui}
\alias{mod_tickets}
\title{mod_ticketsui and mod_tickets}
\usage{
mod_ticketsui(id)

mod_tickets(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that ...
}
